backend/requirements.txt
opencv-python
numpy
tensorflow
flask
pillow
flask-cors 